pyproject.toml
[tool.poetry]
name = "judicor"
version = "0.1.0"
description = "A Python-based incident judgment engine that assists DevOps and SRE teams in real-time production incident analysis using Gemini-powered reasoning with human accountability."
authors = [
    "Adi Roth <adi.build.balance@gmail.com>"
]
license = "MIT"
readme = "README.md"
packages = [
    { include = "judicor", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
typer = ">=0.9.0,<1.0.0"
fastapi = ">=0.110.0,<1.0.0"
uvicorn = ">=0.23.0,<1.0.0"
pydantic = ">=2.0.0,<3.0.0"
google-genai = ">=1.56.0,<2.0.0"
flake8 = "^7.3.0"

[tool.poetry.scripts]
judicor = "judicor.cli.app:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "pytest (>=8.3.0,<9.0.0)",
    "pytest-cov (>=5.0.0,<6.0.0)"
]
